+++ micronaut-mcp-server-java-sdk/src/test/java/io/micronaut/mcp/server/tools/fetch/FetchToolFactoryHttpTest.java
@@ -24,6 +24,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 @Property(name = "micronaut.mcp.server.info.version", value = "0.0.1")
 @Property(name = "micronaut.mcp.server.transport", value = "HTTP")
 @Property(name = "spec.name", value = "FetchToolFactoryTest")
+@Property(name = "spec.tool.name", value = "MicronautModulesFetch")
 @MicronautTest
 class FetchToolFactoryHttpTest {
 
@@ -68,20 +69,4 @@ class FetchToolFactoryHttpTest {
  }""";
         JSONAssert.assertEquals(expected, responseJson, true);
     }
-
-    @Requires(property = "spec.name", value = "FetchToolFactoryTest")
-    @Singleton
-    static class MicronautModulesFetch implements FetchTool {
-        @Override
-        public Optional<FetchResponse> fetch(FetchRequest request, McpTransportContext transportContext) {
-            return Optional.of(FetchResponse.builder()
-                .id("micronaut-security")
-                .title("Micronaut Security")
-                .url("https://micronaut-projects.github.io/micronaut-security/latest/guide")
-                .text("Built-in security features. Authentication providers and strategies, Token Propagation.")
-                .build());
-        }
-    }
-
-
 }
