+++ micronaut-mcp-server-java-sdk/src/test/java/io/micronaut/mcp/server/tools/fetch/FetchToolFactoryStdioAsyncTest.java
@@ -33,6 +33,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 @Property(name = "micronaut.mcp.server.transport", value = "STDIO")
 @Property(name = "micronaut.mcp.server.reactive", value = StringUtils.TRUE)
 @Property(name = "spec.name", value = "FetchToolFactoryStdioAsyncTest")
+@Property(name = "spec.tool.name", value = "MicronautModulesFetch")
 @MicronautTest
 class FetchToolFactoryStdioAsyncTest {
     @Inject
@@ -71,20 +72,6 @@ class FetchToolFactoryStdioAsyncTest {
   }
 }""";
         JSONAssert.assertEquals(expected, responseJson, true);
-    }
-
-    @Requires(property = "spec.name", value = "FetchToolFactoryStdioAsyncTest")
-    @Singleton
-    static class MicronautModulesFetch implements FetchTool {
-        @Override
-        public Optional<FetchResponse> fetch(FetchRequest request, McpTransportContext transportContext) {
-            return Optional.of(FetchResponse.builder()
-                .id("micronaut-security")
-                .title("Micronaut Security")
-                .url("https://micronaut-projects.github.io/micronaut-security/latest/guide")
-                .text("Built-in security features. Authentication providers and strategies, Token Propagation.")
-                .build());
-        }
     }
 
     @Requires(property = "spec.name", value = "FetchToolFactoryStdioAsyncTest")
