+++ micronaut-mcp-server-java-sdk/src/test/java/io/micronaut/mcp/server/tools/search/SearchToolFactoryStdioAsyncTest.java
@@ -32,6 +32,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 @Property(name = "micronaut.mcp.server.transport", value = "STDIO")
 @Property(name = "micronaut.mcp.server.reactive", value = StringUtils.TRUE)
 @Property(name = "spec.name", value = "SearchToolFactoryStdioAsyncTest")
+@Requires(property = "spec.name", value = "MicronautModulesSearch")
 @MicronautTest
 class SearchToolFactoryStdioAsyncTest {
     @Inject
@@ -72,20 +73,6 @@ class SearchToolFactoryStdioAsyncTest {
   }
 }""";
         JSONAssert.assertEquals(expected, responseJson, true);
-    }
-
-    @Requires(property = "spec.name", value = "SearchToolFactoryStdioAsyncTest")
-    @Singleton
-    static class MicronautModulesSearch implements SearchTool {
-
-        @Override
-        public SearchResponse search(SearchRequest request, McpTransportContext transportContext) {
-            return new SearchResponse(List.of(SearchResult.builder()
-                .id("micronaut-security")
-                .title("Micronaut Security")
-                .url("https://micronaut-projects.github.io/micronaut-security/latest/guide")
-                .build()));
-        }
     }
 
     @Requires(property = "spec.name", value = "SearchToolFactoryStdioAsyncTest")
