+++ src/main/docs/guide/server/primitives/tools/fetch.adoc
@@ -1,0 +1,6 @@
+If you need to expose a https://platform.openai.com/docs/mcp#fetch-tool[Fetch Tool] in your MCP Server, you can do it easily defining a bean of type api:mcp.server.tools.fetch.FetchTool[]
+
+[source, Java]
+----
+include::micronaut-mcp-server-java-sdk/src/test/java/io/micronaut/mcp/server/tools/fetch/MicronautModulesFetch.java[tags="clazz",indent=0]
+----
